fix(e2e): rename EnpointPage -> EndpointPage and update export; docs: contributors/architecture folder rename
