+++ packages/e2e/src/page-objects/index.ts
@@ -2,5 +2,5 @@ export { MotiaApplicationPage } from './MotiaApplicationPage'
 export { WorkbenchPage } from './WorkbenchPage'
 export { LogsPage } from './LogsPage'
 export { ApiHelpers } from './ApiHelpers'
-export { EndpointPage } from './EnpointPage'
+export { EndpointPage } from './EndpointPage'
 export { TracesPage } from './TracesPage'
